+++ src/main/java/com/github/afanas10101111/mp/dto/MockRuleTo.java
@@ -10,11 +10,19 @@ import java.util.List;
 
 @Data
 public class MockRuleTo {
-    public static final String STUB_VALIDATION_ERROR_MESSAGE = "stub must be presented";
+    public static final String STATUS_VALIDATION_ERROR_MESSAGE = "status stub must be presented";
+    public static final String CONTENT_TYPE_VALIDATION_ERROR_MESSAGE = "content type must be presented";
+    public static final String BODY_VALIDATION_ERROR_MESSAGE = "body must be presented";
     public static final String PATTERNS_VALIDATION_ERROR_MESSAGE = "at least one pattern must be presented";
 
-    @NotNull(message = STUB_VALIDATION_ERROR_MESSAGE)
-    private String stub;
+    @NotNull(message = STATUS_VALIDATION_ERROR_MESSAGE)
+    private Integer status;
+
+    @NotNull(message = CONTENT_TYPE_VALIDATION_ERROR_MESSAGE)
+    private String contentType;
+
+    @NotNull(message = BODY_VALIDATION_ERROR_MESSAGE)
+    private String body;
 
     @NotEmpty(message = PATTERNS_VALIDATION_ERROR_MESSAGE)
     private List<@Valid PatternKeeperTo> patterns = new ArrayList<>();