+++ src/main/java/com/github/afanas10101111/mp/filter/PreFilter.java
@@ -1,7 +1,8 @@
 package com.github.afanas10101111.mp.filter;
 
-import com.github.afanas10101111.mp.service.RequestBodyChecker;
 import com.github.afanas10101111.mp.config.ProxyConfig;
+import com.github.afanas10101111.mp.model.MockRule;
+import com.github.afanas10101111.mp.service.RequestBodyChecker;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cloud.gateway.filter.GatewayFilterChain;
@@ -9,7 +10,6 @@ import org.springframework.cloud.gateway.filter.GlobalFilter;
 import org.springframework.cloud.gateway.support.ServerWebExchangeUtils;
 import org.springframework.core.io.buffer.DataBuffer;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.server.reactive.ServerHttpResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
@@ -22,7 +22,6 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Component
 public class PreFilter implements GlobalFilter {
-    private static final String CONTENT_TYPE = "text/xml;charset=UTF-8";
     private static final String REQUEST_BODY_OBJECT = "cachedRequestBodyObject";
 
     private final ProxyConfig proxyConfig;
@@ -33,16 +32,16 @@ public class PreFilter implements GlobalFilter {
         String body = exchange.getAttribute(REQUEST_BODY_OBJECT);
         if (body != null) {
             log.info("filter -> request:\n{}", body);
-            Optional<String> stub = checker.getStubbedResponse(body);
-            if (stub.isPresent()) {
-                String stubString = stub.get();
-                log.info("filter -> stubbed response:\n{}", stubString);
-                ServerWebExchangeUtils.setResponseStatus(exchange, HttpStatus.OK);
+            Optional<MockRule> mockRule = checker.getMockRule(body);
+            if (mockRule.isPresent()) {
+                MockRule rule = mockRule.get();
+                log.info("filter -> found rule:\n{}", rule);
+                ServerWebExchangeUtils.setResponseStatus(exchange, rule.getStatus());
                 ServerWebExchangeUtils.setAlreadyRouted(exchange);
                 return chain.filter(exchange).then(Mono.defer(() -> {
                     ServerHttpResponse response = exchange.getResponse();
-                    response.getHeaders().add(HttpHeaders.CONTENT_TYPE, CONTENT_TYPE);
-                    DataBuffer buffer = response.bufferFactory().wrap(stubString.getBytes());
+                    response.getHeaders().add(HttpHeaders.CONTENT_TYPE, rule.getContentType());
+                    DataBuffer buffer = response.bufferFactory().wrap(rule.getBody().getBytes());
                     return response.writeWith(Flux.just(buffer));
                 }));
             }