+++ RELEASE_NOTES.md
@@ -1,7 +1,15 @@
 Release Notes
 ========
 
+## 1.0.0
+
+MVP version is ready:
+* Configure the request body patterns to choose which request should be intercepted
+* Configure the status, content type and body of the response to the intercepted request
+* Configure the repeat counter (how many times will this rule work before forwarding and resetting the counter)
+
 ## 0.1.0-SNAPSHOT
+
 * [FEATURE] MP-4: Deal with mock rules duplication
 * [FEATURE] MP-2: Add the controller for mock rules setup
 * [FEATURE] MP-1: Add the repository for mock rules